e2fsprogs: add patch to avoid getrandom call
This is a workaround for an arbitrarily long hang when mkfs.ext4 is
called to initialize the application filesystem on boot due to the
urandom pool not being initialized.

If the initialization application forces the UUID, no calls to getrandom
will be issued (via uuid_generate) by mkfs.ext4. This is critical to
quickly recovering from a corrupt filesystem.